test: testing the existing code and it works, so far so good

+++ internals/ports/db.go
@@ -4,7 +4,7 @@ import "github.com/somT-oss/urlshortner/internals/application/domain"
 
 
 type DBPort interface {
-	GetUrlById(id int) (domain.ShortenedUrl, error)
+	GetUrlById(id int64) (domain.ShortenedUrl, error)
 	GetUrlByShortCode(shortCode string) (domain.ShortenedUrl, error)
 	SaveShortenedUrl(*domain.ShortenedUrl) error
 }
